Refactor balance
