+++ app/admin/balances.rb
@@ -1,4 +1,4 @@
-ActiveAdmin.register Balance do
+ActiveAdmin.register Account::Balance do
   # See permitted parameters documentation:
   # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
   #
