+++ app/views/account/balances/index.html.erb
@@ -1,0 +1,24 @@
+<div class="title">
+  Balances for <%= @balances[0].account_name%>
+</div>
+
+<table class="table table-hover">
+  <thead>
+    <tr>
+      <th>Balance</th>
+      <th>Date</th>
+      <th colspan="2"></th>
+    </tr>
+  </thead>
+
+  <tbody>
+    <% @balances.limit(12).each do |balance| %>
+      <tr>
+        <td><%= balance.balance %></td>
+        <td><%= balance.date.strftime('%B %d, %Y') %></td>
+        <td><%= link_to 'Edit', edit_account_balance_path(id: balance.id, account_id: balance.account_id), class: 'btn btn-warning' %></td>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
+<br>
