+++ app/views/account/balances/new.html.erb
@@ -1,10 +1,15 @@
-<%= form_for @balance do |form| %>
-  <% if balance.errors.any? %>
+<div class="title">
+New Balance
+</div>
+
+<div class="form">
+<%= form_with(url: account_balances_path, scope: :balance, local:true) do |form| %>
+  <% if @balance.errors.any? %>
     <div id="error_explanation">
-      <h2><%= pluralize(balance.errors.count, "error") %> prohibited this balance from being saved:</h2>
+      <h2><%= pluralize(@balance.errors.count, "error") %> prohibited this balance from being saved:</h2>
 
       <ul>
-        <% balance.errors.full_messages.each do |message| %>
+        <% @balance.errors.full_messages.each do |message| %>
           <li><%= message %></li>
         <% end %>
       </ul>
@@ -23,12 +28,11 @@
     </div>
 
     <div class="form-group">
-      <%= form.label :account_id, class: "input-group-text" %>
-      <%= form.collection_select(:account_id, policy_scope(Account), :id, :name, {}, {class: "custom-select"}) %>
-    </div>
-
-    <div class="form-group">
       <%= form.submit  class: "btn btn-success" %>
     </div>
   </form>
 <% end %>
+
+<%= link_to 'Back', account_path(id: params[:account_id]), class: "btn btn-primary" %>
+
+</div>
