--- app/views/balances/_balance.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! balance, :id, :balance, :date, :account_id, :created_at, :updated_at
-json.url balance_url(balance, format: :json)
