--- app/views/balances/edit.html.erb
@@ -1,9 +1,0 @@
-<div class="title">
-  <%= @balance.account_name %> balance
-</div>
-
-<div class="form">
-<%= render 'form', balance: @balance %>
-
-<%= link_to 'Back', balances_path, class: "btn btn-primary" %>
-</div>
