--- app/views/balances/index.html.erb
@@ -1,34 +1,0 @@
-<div class="title">
-  Balances
-</div>
-
-<% @accounts.each do |account| %>
-
-  <div class="summary-subtitle">
-    <%= link_to account.name, account_path(account) %>
-  </div>
-
-<table class="table table-hover">
-  <thead>
-    <tr>
-      <th>Balance</th>
-      <th>Date</th>
-      <th colspan="2"></th>
-    </tr>
-  </thead>
-
-  <tbody>
-    <% account.balances.order('date desc').limit(5).each do |balance| %>
-      <tr>
-        <td><%= balance.balance %></td>
-        <td><%= balance.date.strftime('%B %d, %Y') %></td>
-        <td><%= link_to 'Edit', edit_account_balance_path(id: balance.id, account_id: balance.account_id), class: 'btn btn-warning' %>
-          <%= link_to 'Delete', account_balance_path(id: balance.id, account_id: balance.account_id), method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-danger' %></td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
-<% end %>
-<br>
-
-<%= link_to 'New Balance', new_balance_path, class: "btn btn-success ml-3" %>
