--- app/views/balances/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @balances, partial: 'balances/balance', as: :balance
