--- app/views/balances/new.html.erb
@@ -1,10 +1,0 @@
-<div class="title">
-New Balance
-</div>
-
-<div class="form">
-<%= render 'form', balance: @balance %>
-
-<%= link_to 'Back', balances_path, class: "btn btn-primary" %>
-
-</div>
