--- app/views/balances/show.html.erb
@@ -1,21 +1,0 @@
-<div class="form-group">
-  <div class="input-group-text">
-    <strong>Account:</strong>
-    <%= @balance.account_name %>
-  </div>
-
-  <div class="input-group-text">
-    <strong>Balance:</strong>
-    <%= @balance.balance %>
-    </div>
-
-  <div class="input-group-text">
-    <strong>Date:</strong>
-    <%= @balance.date.strftime('%B %d, %Y') %>
-  </div>
-
-  <div>
-    <%= link_to 'Edit', edit_account_balance_path(id: @balance.id, account_id: @balance.account_id), :class => "btn btn-warning" %>
-    <%= link_to 'Back', balances_path, :class => "btn btn-primary" %>
-  </div>
-</div>
