--- app/views/balances/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! 'balances/balance', balance: @balance
