+++ config/routes.rb
@@ -8,6 +8,7 @@ Rails.application.routes.draw do
 
   scope module: 'account' do
     resources :accounts do
+      resources :balances, only: %i[index new create]
       resources :transactions, only: %i[index new create edit update]
     end
   end
