+++ spec/models/account/balance_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+RSpec.describe Account::Balance, type: :model do
+  describe 'associations' do
+    it { is_expected.to belong_to(:account) }
+  end
+
+  describe 'monetize' do
+    it { is_expected.to monetize(:balance) }
+  end
+end
